% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCode.R
\name{data_code}
\alias{data_code}
\title{Generate tidyverse code to replicate data edits}
\usage{
data_code(x, x_edit, name = "data")
}
\arguments{
\item{x}{original data prior to editing, a \code{data.frame} or
\code{matrix}. If \code{NULL}, creation code will be generated for
\code{x_edit}.}

\item{x_edit}{edited data after modifications, a \code{data.frame} or
\code{matrix}.}

\item{name}{character string for the name of the data object in the generated
code, set to \code{"data"} by default.}
}
\value{
a character string containing tidyverse code that can be printed to
  the console or written to an R script.
}
\description{
\code{data_code} compares an original dataset with an edited version and
generates tidyverse-style R code (using \code{dplyr} verbs) that transforms
the original data into the edited version. The generated code uses
\code{rename()}, \code{select()}, \code{mutate()}, \code{slice()}, and
\code{tibble::add_row()} to replicate column renames, column
additions/removals, cell value changes, and row additions/removals. When
meaningful row names are present (i.e. non-default row names such as those
in \code{mtcars}), the generated code preserves them using
\code{tibble::rownames_to_column()} and \code{tibble::column_to_rownames()}.
}
\examples{
# original data
x <- data.frame(
  A = c(1, 2, 3),
  B = c("x", "y", "z"),
  stringsAsFactors = FALSE
)

# edited data (changed a value and renamed a column)
x_edit <- data.frame(
  A = c(1, 5, 3),
  Beta = c("x", "y", "z"),
  stringsAsFactors = FALSE
)

# generate code
cat(data_code(x, x_edit, name = "x"))

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
